+++ src/main/java/ru/vyarus/dropwizard/guice/module/jersey/support/LazyGuiceFactory.java
@@ -30,4 +30,9 @@ public class LazyGuiceFactory<T> implements Factory<T> {
     public void dispose(final T instance) {
         // do nothing
     }
+
+    @Override
+    public String toString() {
+        return "LazyGuiceFactory for " + type;
+    }
 }