+++ src/test/groovy/ru/vyarus/dropwizard/guice/resource/ResourcesNonSingletonTest.groovy
@@ -1,32 +1,34 @@
 package ru.vyarus.dropwizard.guice.resource
 
 import ru.vyarus.dropwizard.guice.AbstractTest
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.configurator.GuiceyConfigurator
+import ru.vyarus.dropwizard.guice.module.installer.InstallersOptions
 import ru.vyarus.dropwizard.guice.support.resource.PrototypeResource
 import ru.vyarus.dropwizard.guice.support.resource.ResourceSingletonCheckApplication
 import ru.vyarus.dropwizard.guice.support.resource.SingletonResource
 import ru.vyarus.dropwizard.guice.test.spock.UseDropwizardApp
 
 /**
- * Check resources are singletons
- * @author Vyacheslav Rusakov 
- * @since 04.10.2014
+ * @author Vyacheslav Rusakov
+ * @since 03.05.2018
  */
-@UseDropwizardApp(ResourceSingletonCheckApplication)
-class ResourcesSingletonTest extends AbstractTest {
+@UseDropwizardApp(value = ResourceSingletonCheckApplication, configurators = XConf)
+class ResourcesNonSingletonTest extends AbstractTest {
 
     void cleanupSpec() {
         PrototypeResource.reset()
         SingletonResource.reset()
     }
 
-    def "Check resource singleton"() {
+    def "Check resource non singleton"() {
 
         when: "calling prototype resource"
         new URL("http://localhost:8080/prototype/").getText()
         new URL("http://localhost:8080/prototype/").getText()
         then: "resource instantiated one time, because singletons are forced"
         PrototypeResource.callCounter == 2
-        PrototypeResource.creationCounter == 1
+        PrototypeResource.creationCounter == 2
 
         when: "calling singleton resource"
         new URL("http://localhost:8080/singleton/").getText()
@@ -34,5 +36,12 @@ class ResourcesSingletonTest extends AbstractTest {
         then: "resource instantiated one time"
         SingletonResource.callCounter == 2
         SingletonResource.creationCounter == 1
+    }
+
+    static class XConf implements GuiceyConfigurator {
+        @Override
+        void configure(GuiceBundle.Builder builder) {
+            builder.option(InstallersOptions.ForceSingletonForHkExtensions, false)
+        }
     }
 }