+++ CHANGELOG.md
@@ -36,7 +36,9 @@
         Hk bridge must be enabled (GuiceyOptions.UseHkBridge) for hk-first mode (exception thrown if not).
     - Add @GuiceManaged annotation to mark exceptions in hk-first mode (when @HK2Managed become useless).
        In guice-first mode this annotation is useless.    
-    - Builder shortcut: .useHK2ForJerseyExtensions() to simplify hk-first mode enabling.            
+    - Builder shortcut: .useHK2ForJerseyExtensions() to simplify hk-first mode enabling.
+* Singleton scope is not forced for jersey extensions with explicit scoping annotation
+* Add option for disabling forced singletons for jersey extensions: InstallerOptions.ForceSingletonForHkExtensions                
 
 ### 4.1.0 (2017-05-09)
 * Update to dropwizard 1.1.0